Re-organize around new app design
A track is a collection of stems. Stems are recorded individually and then
mixed together into a single track later on. Stems can be (re)recorded,
played back, added, and removed.

A recording session manager handles the creation of the current track and
its stems. The presenter interfaces with the recording session manager and
provides the view with its results.