Set up PCM recording
The `PcmRecorderRunnable` abstract class is a general utility class for recording
PCM audio. It provides three abstract methods that serve as hooks into the
recording process.

- `onRecordStart()` is executed before the recording loop begins
- `onRecord()` is executed during the recording loop
- `onRecordStop()` is executed after the recording stops

The `StemRecorder` class extends `PcmRecorderRunnable` and defines how a
stem is recorded. It will also handle the writing of PCM data to file
later on.