Fix StemRecorder post-record buffer preparation
`onRecordStop()` incorrectly assumed that the provided method argument
signified the number of recorded bytes, instead of shorts. This led to the
actual audio buffer being cut in half because only every second recorded
sample was stored in the final buffer. This caused a "speedup" of the
recording.

The "bytes" were also mistakenly combined into shorts, causing clipping
noise.